codeforces global round 10

+++ cgr10-E.cpp
@@ -7,7 +7,7 @@ void make_arr() {
     for(int i = 0; i < n; i++){
         for(int j = 0; j < n; j++){
             if(i & 1){
-                arr[i][j] = (ll)1 << (i + j);
+                arr[i][j] = 1LL << (i + j);
             }else{
                 arr[i][j] = 0;
             }
@@ -37,10 +37,10 @@ int main() {
         cout << point;
         for(int i = 1; i <= (n - 1) * 2; i++){
             if(point.first & 1){//单数层
-                if(g & ((ll)1 << i))    point.first++;
+                if(g & (1LL << i))    point.first++;
                 else    point.second++;
             }else{
-                if(g & ((ll)1 << i))    point.second++;
+                if(g & (1LL << i))    point.second++;
                 else    point.first++;
             }
             cout << point;
